Restrucutre files and directories
common dir: holds assets, scripts, components, styles, etc that are common to the application. 
dist: production ready files
lib: vendor files
topics: for this project, it is important to showcase the topics that are covered in the AngularJS Developer Guide clearly. Each topics dir holds a specific category covered in the guide.

Within each directory, when the number of files get too large, it is ideal to divide them in components, assets, styles, scripts, partials, etc. Each topic should be seen as a sub-app.